Allows the ability to move wirecutters
By taking the components of the wirecutter, creating an empty, and putting each component in the empty, the wirecutters start acting like a cohesive object. Next you add a box collider around the object, mine had a center of 1,4.9,-3.5, and a size of 6,5,2.  I had to use a box collider since the wirecutter isnt just 1 object, although I might be able to go in blender scale it up by 1.1, make it all one object to use it as a mesh collider? then you attach the script to it, and it should work. It is not changing position on the z axis, but that is a problem for another day